+++ common/src/main/java/de/maxhenkel/voicechat/config/ClientConfig.java
@@ -30,5 +30,6 @@ public abstract class ClientConfig {
     public ConfigEntry<Double> hudIconScale;
     public ConfigEntry<String> recordingDestination;
     public ConfigEntry<Boolean> denoiser;
+    public ConfigEntry<Boolean> soundPhysics;
 
 }