+++ fabric/src/main/java/de/maxhenkel/voicechat/config/FabricClientConfig.java
@@ -31,6 +31,7 @@ public class FabricClientConfig extends ClientConfig {
         hudIconScale = builder.doubleEntry("hud_icon_scale", 1D, 0.01D, 10D);
         recordingDestination = builder.stringEntry("recording_destination", "");
         denoiser = builder.booleanEntry("denoiser", false);
+        soundPhysics = builder.booleanEntry("soundphysics", true);
     }
 
 }