+++ fabric/src/main/java/de/maxhenkel/voicechat/intercompatibility/FabricCommonCompatibilityManager.java
@@ -117,4 +117,9 @@ public class FabricCommonCompatibilityManager extends CommonCompatibilityManager
         return sb.toString();
     }
 
+    @Override
+    public boolean isDevEnvironment() {
+        return FabricLoader.getInstance().isDevelopmentEnvironment();
+    }
+
 }