+++ forge/src/main/java/de/maxhenkel/voicechat/intercompatibility/ForgeClientCompatibilityManager.java
@@ -3,7 +3,10 @@ package de.maxhenkel.voicechat.intercompatibility;
 import com.mojang.blaze3d.platform.InputConstants;
 import de.maxhenkel.voicechat.events.VoiceChatConnectedEvent;
 import de.maxhenkel.voicechat.events.VoiceChatDisconnectedEvent;
+import de.maxhenkel.voicechat.voice.client.ALSpeaker;
 import de.maxhenkel.voicechat.voice.client.ClientVoicechatConnection;
+import de.maxhenkel.voicechat.voice.client.SoundManager;
+import de.maxhenkel.voicechat.voice.client.SpeakerException;
 import net.minecraft.client.KeyMapping;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.sounds.SoundEngineExecutor;
@@ -19,6 +22,7 @@ import net.minecraftforge.event.world.WorldEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fmlclient.registry.ClientRegistry;
 
+import javax.annotation.Nullable;
 import java.net.SocketAddress;
 import java.util.ArrayList;
 import java.util.List;
@@ -170,5 +174,17 @@ public class ForgeClientCompatibilityManager extends ClientCompatibilityManager 
     @Override
     public void addResourcePackSource(PackRepository packRepository, RepositorySource repositorySource) {
         packRepository.addPackFinder(repositorySource);
+    }
+
+    @Override
+    public ALSpeaker createSpeaker(SoundManager soundManager, int sampleRate, int bufferSize) {
+        return new ALSpeaker(soundManager, sampleRate, bufferSize) {
+        };
+    }
+
+    @Override
+    public SoundManager createSoundManager(@Nullable String deviceName) throws SpeakerException {
+        return new SoundManager(deviceName) {
+        };
     }
 }